begin rewrite
