+++ gui/Cargo.toml
@@ -1,23 +1,19 @@
 [package]
 name = "gui"
 version = "0.1.0"
-authors = ["MinusKelvin <minuskelvin.carlson@gmail.com>"]
+authors = ["MinusKelvin <mark.carlson@minuskelvin.net>"]
 edition = "2018"
 default-run = "gui"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-arrayvec = "0.4.11"
-ggez = "0.5.1"
+game-util = { git = "https://github.com/MinusKelvin/game-util-rs" }
+cold-clear = { path = "../bot" }
 libtetris = { path = "../libtetris" }
 battle = { path = "../battle" }
-cold-clear = { path = "../bot" }
-serde = "1"
-bincode = "1"
-libflate = "0.1"
-serde_yaml = "0.8"
-rand = "0.7.0"
-rand_pcg = "0.2.0"
-enumset = "0.4.0"
-winit = { version = "0.19", features = ["serde"] }
+arrayvec = "0.4.11"
+rand = "0.7.3"
+
+[build-dependencies]
+build-utils = { git = "https://github.com/MinusKelvin/game-util-rs" }
